TradingBot/requirements.txt
requests~=2.32.3
sortedcontainers~=2.4.0
PyYAML~=6.0.2
dash-bootstrap-components
dash~=2.18.1
websockets~=13.1
plotly~=5.24.1
pandas~=2.2.3
binance-connector~=3.5.1